requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls
telethon==1.28.2
pyrogram==2.0.106
youtube-search
youtube-search-python
yt-dlp
python-dotenv
dnspython
gitpython
asgiref
aiofiles
aiohttp
pytz
requests
pillow
heroku3
psutil
future
wget
speedtest-cli
motor
pyromod~=1.5
googletrans
pytz
